chore:컴포넌트 디렉터리 구조 일부 변경
QueryClientProvider, RecoilRoot 모두 index 파일에서 사용
컴포넌트 파일에서 다시 세부파일로 나누지 않도록 수정
