+++ src/App.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { lazy, Suspense, useEffect, useState } from 'react';
 import { BrowserRouter, Router, Route, Routes } from 'react-router-dom';
-import MainPage from './pages/MainPage/MainPage';
+import MainPage from './pages/MainPage';
 import ModalProvider from './contexts/ModalContext';
 import KakaoAuthPage from './pages/KakaoAuthPage';
 import NaverAuthPage from './pages/NaverAuthPage';
