+++ src/components/@icons/index.ts
@@ -1,3 +1,3 @@
 export { default as UserIcon } from './User';
 export { default as GpsIcon } from './Gps';
-export { techs} from './techs/techs';
+export { techs} from './techs/Techs';
