+++ src/components/Study/StudyContainer.style.ts
@@ -1,7 +1,16 @@
 import styled from '@emotion/styled';
 import Flex from 'src/styles/Flex';
 
-
+const Main = styled.main`
+position:relative;
+top: 4rem;
+width:1180px;
+min-height:50rem;
+display:flex;
+Flex-direction:column;
+margin:0 auto;
+z-index:-1;
+`
 
 
 const StudySection = styled.section`
@@ -14,4 +23,6 @@ const StudySection = styled.section`
 
 
 
-export { StudySection }
+
+
+export { StudySection, Main }
