+++ src/components/Study/StudyContainer.tsx
@@ -1,6 +1,6 @@
-import { StudySection } from './StudyContainer.style';
+import { StudySection, Main } from './StudyContainer.style';
 import { useRecoilValue } from 'recoil';
-import StudyList from './StudyList/StudyList';
+import StudyList from './StudyList';
 import getStudy from '../../apis/studyTest';
 import { useEffect } from 'react';
 import { useQuery } from 'react-query';
@@ -10,18 +10,14 @@ import axios from 'axios';
 
 
 const StudyListContainer = () => {
-/*   const { isLoading, error, data, isFetching } = useQuery(
-    ["StudyList"],
-    () => {
-      
-    } ) */
-
 
   
   return (
-    <StudySection>
-      <StudyList />
-    </StudySection>
+    <Main>
+      <StudySection>
+        <StudyList />
+      </StudySection>
+    </Main>
   );
 }
