--- src/pages/MainPage/MainPage.style.ts
@@ -1,15 +1,0 @@
-import styled from '@emotion/styled';
-
-
-const Main = styled.main`
-position:relative;
-top: 4rem;
-width:1180px;
-min-height:50rem;
-display:flex;
-Flex-direction:column;
-margin:0 auto;
-z-index:-1;
-`
-
-export { Main }
