--- src/pages/MainPage/MainPage.tsx
@@ -1,24 +1,0 @@
-import HeaderNavigation from "src/components/Header/HeaderNavigation";
-import { QueryClient, QueryClientProvider } from "react-query";
-import StudyListContainer from "src/components/Study/StudyContainer";
-import Statistics from "../../components/Statistics/Statistics";
-import { Main } from './MainPage.style';
-
-
-const queryClient = new QueryClient();
-
-const MainPage = () => {
-  return (
-    <>
-      <QueryClientProvider client={queryClient}>
-      <HeaderNavigation />
-      <Statistics />
-      <Main>
-        <StudyListContainer />  
-      </Main>
-      </QueryClientProvider>
-    </>
-  );
-}
-
-export default MainPage;
